chore: move SheetsHomePage into sheets directory
