+++ static/js/ReaderApp.jsx
@@ -36,7 +36,7 @@ import { Promotions } from './Promotions';
 import Component from 'react-class';
 import  { io }  from 'socket.io-client';
 import { SignUpModalKind } from './sefaria/signupModalContent';
-import SheetsHomePage from "./SheetsHomePage";
+import SheetsHomePage from "./Sheets/SheetsHomePage";
 
 class ReaderApp extends Component {
   constructor(props) {
